Cast AP matrix seed values for remote execution
The linked query runner inferred nullable matrix seed columns as text, which failed against numeric route bounds. The seed now casts min/max amount values explicitly so remote execution matches the table shape.

Constraint: Supabase Management API query execution can infer VALUES types differently from local static checks.
Rejected: Remove max_amount/min_amount from the seed | route bounds are part of the AP approval matrix contract.
Confidence: high
Scope-risk: narrow
Directive: Cast nullable VALUES tuples when the target column type matters.
Tested: bun run migrations:check
Not-tested: Remote retry beyond the next targeted apply attempt.

+++ supabase/migrations/658_finance_foundation_ap_three_way_match.sql
@@ -325,9 +325,9 @@ insert into public.ap_approval_matrix (
 select *
 from (
   values
-  ('default', 'ap_standard_finance_admin', 'Standard AP invoice finance review', 0, null, 'finance_admin', 1, true),
-  ('default', 'ap_owner_exception', 'Owner review for AP exception invoices and unmatched holds', null, null, 'owner', 2, true),
-  ('default', 'ap_reconditioning_soft_cap', 'Reconditioning soft-cap route; inactive until threshold is configured', null, null, 'manager', 1, false)
+  ('default', 'ap_standard_finance_admin', 'Standard AP invoice finance review', 0::numeric, null::numeric, 'finance_admin', 1, true),
+  ('default', 'ap_owner_exception', 'Owner review for AP exception invoices and unmatched holds', null::numeric, null::numeric, 'owner', 2, true),
+  ('default', 'ap_reconditioning_soft_cap', 'Reconditioning soft-cap route; inactive until threshold is configured', null::numeric, null::numeric, 'manager', 1, false)
 ) as seed(workspace_id, route_code, description, min_amount, max_amount, required_role, approval_sequence, active)
 where not exists (
   select 1
